External QR Report Views.
creation fro the related files

+++ modules/reporting/sendbymail.php
@@ -29,7 +29,6 @@ if(!$core->input['action']) {
 		/* 	list($suppliername, $eid) = $db->fetch_array($db->query("SELECT e.companyName AS suppliername, e.eid FROM ".Tprefix."entities e, ".Tprefix."reports r 
 		  WHERE r.spid=e.eid AND r.rid='".$db->escape_string($meta_data['spid'][1])."'"));
 		 */
-	
 		switch($meta_data['type']) {
 			case 'm': $type = 'monthly';
 				$subject_monthquarter = $lang->{strtolower(date('F', mktime(0, 0, 0, $meta_data['month'], 1, 0)))};
@@ -72,6 +71,7 @@ if(!$core->input['action']) {
 				$representatives_list .= '<tr><td style="width: 20%; font-weight: bold;">';
 				$representatives_list .= '<input type="checkbox" name="recipients[]" id="recipient_'.$representative['rpid'].'" value="'.base64_encode($representative['email']).'" checked/> '.$representative['name'].'</td>';
 				$representatives_list .= '<td>'.$representative['email'].'</td></tr>';
+				$hiddenfield.='<td><input type="hidden" name="recipientsid[]" value='.$representative['rpid'].'></td>';
 			}
 		}
 		//$default_cc = $core->settings['sendreportsto'];
@@ -104,6 +104,7 @@ if(!$core->input['action']) {
 }
 else {
 	if($core->input['action'] == 'do_sendbymail') {
+		$meta_data = unserialize($session->get_phpsession('reportmeta_'.$db->escape_string($core->input['identifier'])));
 		if(empty($core->input['recipients']) && empty($core->input['additional_recipients'])) {
 			error($lang->norecipientsselected, $_SERVER['HTTP_REFERER']);
 		}
@@ -113,7 +114,7 @@ else {
 		}
 
 		if(is_array($core->input['recipients'])) {
-			foreach($core->input['recipients'] as $val) {
+			foreach($core->input['recipients'] as $recid => $val) {
 				$email = base64_decode($val);
 				if(isvalid_email($email)) {
 					$valid_emails[] = $email;
@@ -122,11 +123,21 @@ else {
 					$bad_emails[] = $email;
 				}
 			}
+			switch($meta_data['type']) {
+				case'q':
+					/*  recorded in a recipients table --START */
+					$report = new ReportingQr($meta_data);
+					foreach($core->input[recipientsid] as $rpid) {
+						$report->create_recipients($rpid, $core->input['identifier']);
+					}
+
+					/*  recorded in a recipients table --END */
+					break;
+			}
 		}
 
 		if(!empty($core->input['additional_recipients'])) {
 			$additional_emails = explode(',', $core->input['additional_recipients']);
-
 			foreach($additional_emails as $val) {
 				if(isvalid_email(trim($val))) {
 					$cc_valid_emails[] = trim($val);
@@ -136,36 +147,66 @@ else {
 				}
 			}
 		}
-
 		$core->input['message'] = $core->sanitize_inputs($core->input['message'], array('method' => 'striponly', 'allowable_tags' => '<span><div><a><br><p><b><i><del><strike><img><video><audio><embed><param><blockquote><mark><cite><small><ul><ol><li><hr><dl><dt><dd><sup><sub><big><pre><figure><figcaption><strong><em><table><tr><td><th><tbody><thead><tfoot><h1><h2><h3><h4><h5><h6>', 'removetags' => true));
+		switch($meta_data['type']) {
+			case'm':
+				$email_data = array(
+						'from_email' => 'reporting@ocos.orkila.com',
+						'from' => 'Orkila Reporting System',
+						'to' => $valid_emails,
+						'cc' => $cc_valid_emails,
+						'subject' => $core->input['subject'],
+						'message' => $core->input['message'],
+						'attachments' => array($core->input['attachment'])
+				);
+				break;
+		}
 
-		$email_data = array(
-				'from_email' => 'reporting@ocos.orkila.com',
-				'from' => 'Orkila Reporting System',
-				'to' => $valid_emails,
-				'cc' => $cc_valid_emails,
-				'subject' => $core->input['subject'],
-				'message' => $core->input['message'],
-				'attachments' => array($core->input['attachment'])
-		);
 
-		$mail = new Mailer($email_data, 'php');
-		$meta_data = unserialize($session->get_phpsession('reportsmetadata_'.$db->escape_string($core->input['identifier'])));
-		if($mail->get_status() === true) {
-			if(is_array($meta_data['rid'])) {
-				$update_query_where = 'rid IN ('.implode(',', $meta_data['rid']).')';
+		switch($meta_data['type']) {
+			case'q':
+				$recipient = $report->get_recipient($core->input[recipientsid]);
+				if(is_array($recipient)) {
+					foreach($recipient as $rcid => $recipentdata) {
+						$previewreport_link = '';
+						$body_message = '';
+						$previewreport_link[$rcid] = 'http://www.orkila.com/reporting/preview&reportidentifier='.$recipentdata['reportidentifier'].'&token='.$recipentdata['token'].'';
+						$body_message[$rcid] = '<br>'.$previewreport_link[$rcid].' <br>Your new  passowrd is:"'.$recipentdata['password'].'"<br>';
+						$recipientemail_data[$rcid] = array(
+								'from_email' => 'reporting@ocos.orkila.com',
+								'from' => 'Orkila Reporting System',
+								'to' => $recipentdata['email'],
+								'subject' => 'External link to view quarterly Report',
+								'message' => 'follow the blow link :<br>'.$body_message[$rcid],
+						);
+						$mail = new Mailer($recipientemail_data[$rcid], 'php');
+					}
+				}
+				redirect('index.php?module=reporting/generatereport', 1, $lang->messagesentsuccessfully);
+				break;
+		}
+		if(is_array($email_data)) {
+			$mail = new Mailer($email_data, 'php');
+
+			if($mail->get_status() === true) {
+				if(is_array($meta_data['rid'])) {
+					$update_query_where = 'rid IN ('.implode(',', $meta_data['rid']).')';
+				}
+				else {
+					$update_query_where = "rid = '{$meta_data[rid]}'";
+				}
+				$db->update_query('reports', array('isSent' => 1, 'isApproved' => 1, 'isLocked' => 1), $update_query_where);
+
+				log_action($valid_emails, $cc_valid_email);
+				if(is_array($core->input['attachment'])) {
+					unlink($core->input['attachment']);
+				}
+				exit;
+				redirect('index.php?module=reporting/generatereport', 1, $lang->messagesentsuccessfully);
 			}
 			else {
-				$update_query_where = "rid = '{$meta_data[rid]}'";
+				error($lang->errorsendingemail);
 			}
-			$db->update_query('reports', array('isSent' => 1, 'isApproved' => 1, 'isLocked' => 1), $update_query_where);
-
-			log_action($valid_emails, $cc_valid_email);
-			unlink($core->input['attachment']);
-			redirect('index.php?module=reporting/generatereport', 1, $lang->messagesentsuccessfully);
-		}
-		else {
-			error($lang->errorsendingemail);
 		}
 	}
 }
